refactor(spf)!: own the SPF-backed Medias, inverting the media dependency
SPF is unlike hls.js or dash.js in that it publishes no engine-shaped API
for consumers to drive: the `@videojs/media` facade *is* how a composed
engine is consumed. The code locations didn't reflect that — the adapters
lived here while the Medias they compose lived in `@videojs/media`, which
depended on this package.

Invert it. `SimpleHlsMedia`, `SimpleHlsAudioOnlyMedia`, and the media-tracks
projection move into `playback/adapters/`, `@videojs/spf` gains
`@videojs/media`, and `@videojs/media` no longer imports spf at all. With the
facade and the engine in one package, a new Media capability that reads engine
state no longer has to promote engine internals to public API to reach a
consumer across a package boundary — the alternative direction needed 21 such
symbols exported for exactly that reason.

Two compatibility shims that only existed because spf sat below media are
gone: `SimpleHlsMediaError` now extends `ErrorLike`, and
`SimpleHlsMediaStreamType` *is* `MediaStreamType` rather than a
structurally-compatible copy.

The Medias ship behind their own entry points, `@videojs/spf/simple-hls` and
`@videojs/spf/simple-hls-audio-only`, so driving an engine directly pulls in
neither a Media nor `@videojs/media`. That also keeps `./hls` measurable as
the engine's own size budget, taking it from 20.40 KB to 19.08 KB gzipped —
back under its 20 KB target, which it had been exceeding at 102%.

`@videojs/media` exposes `./media-tracks` for the track-list infrastructure
the video Media composes. This is a breaking change for anything importing
the Medias from `@videojs/media/dom/simple-hls*`; html and react move to the
new paths.

`background-video`'s adapter stays in `playback/engines/` for now — it has no
Media and no consumer outside this package's tests.
