+++ packages/html/src/media/simple-hls-audio-only/media.ts
@@ -1,5 +1,5 @@
 import { CustomMediaElement } from '@videojs/media/dom/custom-media-element';
-import { SimpleHlsAudioOnlyMedia } from '@videojs/media/dom/simple-hls-audio-only';
+import { SimpleHlsAudioOnlyMedia } from '@videojs/spf/simple-hls-audio-only';
 import { MediaAttachMixin } from '../../store/media-attach-mixin';
 
 export class SimpleHlsAudioOnly extends MediaAttachMixin(CustomMediaElement('audio', SimpleHlsAudioOnlyMedia)) {}
