--- packages/media/src/dom/simple-hls-audio-only/index.ts
@@ -1,1 +1,0 @@
-export * from './media';
