--- packages/media/src/dom/simple-hls/index.ts
@@ -1,1 +1,0 @@
-export * from './media';
