+++ packages/spf/src/playback/adapters/simple-hls/media.ts
@@ -1,6 +1,6 @@
-import { SimpleHlsMediaMixin } from '@videojs/spf/hls';
-import { MediaTracksMixin } from '../../core/media-tracks';
-import { HTMLVideoElementHost } from '../video-host';
+import { HTMLVideoElementHost } from '@videojs/media/dom/video-host';
+import { MediaTracksMixin } from '@videojs/media/media-tracks';
+import { SimpleHlsMediaMixin } from './adapter';
 import { SimpleHlsMediaMediaTracksMixin } from './media-tracks';
 
 const SimpleHlsMediaBase = SimpleHlsMediaMediaTracksMixin(MediaTracksMixin(SimpleHlsMediaMixin(HTMLVideoElementHost)));
