+++ packages/spf/src/playback/adapters/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "extends": "../../../../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true,
+    "lib": ["ES2022", "DOM", "DOM.Iterable"],
+    "exactOptionalPropertyTypes": false,
+    "declarationDir": "../../../types/playback/adapters"
+  },
+  "references": [
+    { "path": "../../../../utils" },
+    { "path": "../../../../media" },
+    { "path": "../../core" },
+    { "path": "../../network" },
+    { "path": "../../media" },
+    { "path": "../../media/dom" },
+    { "path": "../primitives" },
+    { "path": "../behaviors" },
+    { "path": "../behaviors/dom" },
+    { "path": "../actors" },
+    { "path": "../actors/dom" },
+    { "path": "../engines/hls" }
+  ],
+  "include": ["./**/*.ts"]
+}
