+++ packages/spf/tsdown.config.ts
@@ -10,6 +10,8 @@ const createConfig = (mode: PackageBuildMode): UserConfig => ({
     hls: 'src/playback/engines/hls/index.ts',
     'media-tracks': 'src/media/media-tracks/index.ts',
     'background-video': 'src/playback/engines/background-video/index.ts',
+    'simple-hls': 'src/playback/adapters/simple-hls/index.ts',
+    'simple-hls-audio-only': 'src/playback/adapters/simple-hls-audio-only/index.ts',
   },
 });
